Addressing errors: NaN values in true data and y_training set not being flattened
I added a preprocessing stage to each file to drop
rows with incomplete data then saved the before and
after row count to a txt file labeled with the
respective task. In the working notebook I adjusted the
PseudoData to include NaN records and tested the outlined
implementations.

I also added a function to flatten the y_training values.